Add crop water

+++ src/UIText.cpp
@@ -7,7 +7,7 @@
 #include "Log.h"
 
 UIText::UIText(float pX, float pY, const std::string& pText, const freetype::font_data& pFont, eJustify pJustify)
-	: UIFrame(pX, pY, pFont.width(pText), pFont.h)
+	: UIFrame(pX, pY, (float)pFont.width(pText), (float)pFont.h)
 	, font(pFont)
 	, mText(pText)
 	, justify(pJustify)
@@ -16,7 +16,7 @@ UIText::UIText(float pX, float pY, const std::string& pText, const freetype::fon
 }
 
 UIText::UIText(float pX, float pY, const freetype::font_data& pFont, eJustify pJustify)
-	: UIFrame(pX, pY, 0, pFont.h)
+	: UIFrame(pX, pY, 0, (float)pFont.h)
 	, font(pFont)
 	, mText("")
 	, justify(pJustify)
@@ -27,7 +27,7 @@ UIText::UIText(float pX, float pY, const freetype::font_data& pFont, eJustify pJ
 void UIText::setText(const std::string& pText)
 {
 	mText = pText;
-	width(font.width(mText));
+	width((const float)font.width(mText));
 }
 
 void UIText::setJustify(eJustify j)
